+++ rpc/ethereum/namespaces/eth/filters/utils.go
@@ -57,9 +57,10 @@ func includes(addresses []common.Address, a common.Address) bool {
 	return false
 }
 
+// https://github.com/ethereum/go-ethereum/blob/v1.10.14/eth/filters/filter.go#L321
 func bloomFilter(bloom ethtypes.Bloom, addresses []common.Address, topics [][]common.Hash) bool {
-	var included bool
 	if len(addresses) > 0 {
+		var included bool
 		for _, addr := range addresses {
 			if ethtypes.BloomLookup(bloom, addr) {
 				included = true
@@ -72,15 +73,18 @@ func bloomFilter(bloom ethtypes.Bloom, addresses []common.Address, topics [][]co
 	}
 
 	for _, sub := range topics {
-		included = len(sub) == 0 // empty rule set == wildcard
+		included := len(sub) == 0 // empty rule set == wildcard
 		for _, topic := range sub {
 			if ethtypes.BloomLookup(bloom, topic) {
 				included = true
 				break
 			}
 		}
+		if !included {
+			return false
+		}
 	}
-	return included
+	return true
 }
 
 // returnHashes is a helper that will return an empty hash array case the given hash array is nil,